Create Case Study:
Study how the beak depth (the distance, top to bottom, of a closed beak) of the finch species Geospiza Scandens has changed over time. The Grants have noticed some changes in beak geometry depending on the types of seeds available on the island, and they also noticed that there was some interbreeding with another major species on Daphne Major, Geospiza fortis. These effects can lead to changes in the species over time.